==> Data_preparation/2-ICM_creation.ipynb <==
import pandas as pd
import os
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

base_path = '/content/drive/MyDrive/RS_final_project/recsys2022-main/dataset'

original_data = os.path.join(base_path, 'original_data')
processed_data = os.path.join(base_path, 'processed_data')

features=pd.read_csv(os.path.join(processed_data, "item_features_mapped.csv"))
features=features[features["feature_category_id"]!=27].copy()#useless category with only one item -> no information to gain

features["value"]=1

len(features)

features.head()

features.info()

feat_categories=features["feature_category_id"].unique()
print("min:",min(feat_categories),"  max:",max(feat_categories))
print("length:", len(feat_categories))

feat_values=features["feature_value_id"].unique()
print("min:",min(feat_values),"  max:",max(feat_values))
print("length:", len(feat_values))

feat_category_count=features.groupby("feature_category_id")['value'].count()
feat_category_count # In how many items the feature is present

print("min:",min(feat_category_count),"  max:",max(feat_category_count))
print("length:", len(feat_category_count))

feat_category_count[feat_category_count<500] #features that appears in less than 500 items

feat_category_count[feat_category_count>20000] # features that appears in more than 20000 items

features_category=features[["item_id","feature_category_id"]].copy()

feat_value_count=features.groupby("feature_value_id")['value'].count()
feat_value_count # In how many items the feature is present

print("min:",min(feat_value_count),"  max:",max(feat_value_count))
print("length:", len(feat_value_count))

feat_value_count[feat_value_count<20]

feat_value_count[feat_value_count>20000] # features that appears in more than 20000 items

feats=features.groupby(["feature_category_id","feature_value_id"])["value"].count().reset_index()
feats # For the given feature category, how many items have the value specified in feature_value_id?

feats=feats.sort_values(by="value",ascending=False)
feats
feats.head()

len(feats)

feats.tail()

len(feats[feats["value"]<30])

feats=feats[feats["value"]>=30]

feats=feats.reset_index()
feats["feature_idx"]=feats.index

feats

feats[["feature_category_id","feature_value_id","feature_idx"]].to_csv(os.path.join(processed_data,"feature_mapped_30.csv"), index=False)

feats

feats_map = feats[["feature_category_id","feature_value_id","feature_idx"]]

feats_map

features_mapped=features.merge(feats_map,how="left",on=["feature_category_id","feature_value_id"])

features_category=features_mapped[["item_id","feature_category_id"]].copy()
features_category=features_category.drop_duplicates()

features_category=features_category.rename(columns={"feature_category_id":"feature_idx"})
features_category["feature_idx"]+=np.max(features_mapped["feature_idx"])
features_category.head()


features_category["feature_idx"]=features_category["feature_idx"].astype("int32")

features_mapped=features_mapped[["item_id","feature_idx"]]

features_mapped.info()

features_category

features_mapped.info()

features_mapped.dropna(inplace=True)

features_mapped=features_mapped[["item_id","feature_idx"]]
features_mapped["feature_idx"]=features_mapped["feature_idx"].astype("int32")

features_mapped.info()

features_mapped.to_csv(os.path.join(processed_data,"simplified_features_30.csv"),index=False)

pd.concat([features_mapped,features_category]).to_csv(os.path.join(processed_data,"simplified_features_and_categories_30.csv"),index=False)

pd.concat([features_mapped,features_category])
